requirements.txt
pytest>=3.0.7
Flask>=2.0.2
flask-cors>=3.0.10
manga-ocr>=0.1.4
Pillow~=9.0.0
opencv-python~=4.5.5.62
np~=1.0.2